Cleanup and a handful of docs
