+++ src/Xamarin.Android.Build.Tasks/Tasks/GenerateCompressedAssembliesNativeSourceFiles.cs
@@ -74,22 +74,12 @@ namespace Xamarin.Android.Tasks
 
 			void Generate (IDictionary<string, CompressedAssemblyInfo> dict)
 			{
-				var llvmAsmgen = new LlvmCompressedAssembliesNativeAssemblyGenerator (dict);
+				var llvmAsmgen = new CompressedAssembliesNativeAssemblyGenerator (dict);
 				llvmAsmgen.Init ();
 
 				foreach (string abi in SupportedAbis) {
 					string baseAsmFilePath = Path.Combine (EnvironmentOutputDirectory, $"compressed_assemblies.{abi.ToLowerInvariant ()}");
-					string asmFilePath = $"{baseAsmFilePath}.s";
 					string llvmIrFilePath = $"{baseAsmFilePath}.ll";
-					var asmgen = new CompressedAssembliesNativeAssemblyGenerator (GeneratePackageManagerJava.GetAndroidTargetArchForAbi (abi), dict, baseAsmFilePath);
-
-					using (var sw = MemoryStreamPool.Shared.CreateStreamWriter ()) {
-						asmgen.Write (sw, asmFilePath);
-						sw.Flush ();
-						if (Files.CopyIfStreamChanged (sw.BaseStream, asmFilePath)) {
-							Log.LogDebugMessage ($"File {asmFilePath} was regenerated");
-						}
-					}
 
 					using (var sw = MemoryStreamPool.Shared.CreateStreamWriter ()) {
 						llvmAsmgen.Write (GeneratePackageManagerJava.GetAndroidTargetArchForAbi (abi), sw, llvmIrFilePath);
