--- src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/SectionType.cs
@@ -1,14 +1,0 @@
-namespace Xamarin.Android.Tasks
-{
-	enum SectionType
-	{
-		None,
-		Data, // @progbits
-		NoData, // @nobits
-		InitArray, // @init_array
-		FiniArray, // @fini_array
-		PreInitArray, // @preinit_array
-		Number, // @<number>
-		Custom, // @<target specific>
-	}
-}
