--- src/Xamarin.Android.Build.Tasks/Utilities/NativeAssemblyGenerator/SymbolType.cs
@@ -1,13 +1,0 @@
-namespace Xamarin.Android.Tasks
-{
-	enum SymbolType
-	{
-		Function,
-		GnuIndirectFunction,
-		GnuUniqueObject,
-		Object,
-		TlsObject,
-		Common,
-		NoType,
-	}
-}
